Refactor and Improve `Thumb`
