+++ src/components/Panel3.tsx
@@ -8,14 +8,15 @@ import Animated, {
   useSharedValue,
   withTiming,
 } from 'react-native-reanimated';
+
+import Thumb from './Thumb/Thumb';
+import { CTX } from '../ColorPicker';
 import { clamp } from '../utils';
-import Thumb from './Thumbs';
+import { styles } from '../styles';
 
 import type { LayoutChangeEvent } from 'react-native';
 import type { PanelProps } from '../types';
 import type { PanGestureHandlerEventPayload } from 'react-native-gesture-handler';
-import { CTX } from '../ColorPicker';
-import { styles } from '../styles';
 
 export function Panel3({ thumbShape, thumbSize, thumbColor, renderThumb, thumbStyle, thumbInnerStyle, style = {} }: PanelProps) {
   const {
