+++ src/components/Thumb/BuiltinThumbs/index.ts
@@ -1,0 +1,27 @@
+import Ring from './Ring';
+import Solid from './Solid';
+import Hollow from './Hollow';
+import Line from './Line';
+import Pill from './Pill';
+import Plus from './Plus';
+import TriangleUp from './TriangleUp';
+import TriangleDown from './TriangleDown';
+import DoubleTriangle from './DoubleTriangle';
+import Rect from './Rect';
+import Circle from './Circle';
+
+const BuiltinThumbs = {
+  Ring,
+  Solid,
+  Hollow,
+  Line,
+  Pill,
+  Plus,
+  TriangleUp,
+  TriangleDown,
+  DoubleTriangle,
+  Rect,
+  Circle,
+};
+
+export default BuiltinThumbs;
